+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- added new subcommand `import`, which at this stage converts fastNLO tables to
+  PineAPPL grids. This program was previously an example and written in C++ and
+  now has been removed.
+
 ### Changed
 
 - when running `pineappl convolute ... -s 1` the scale-variation columns are no